+++ modules/unitframes/Defaults.lua
@@ -34,6 +34,7 @@ NivUI.UnitFrames.FRAME_TYPES = {
     { value = "targettarget", name = "Target of Target" },
     { value = "party", name = "Party" },
     { value = "boss", name = "Boss" },
+    { value = "arena", name = "Arena" },
     { value = "raid10", name = "Raid (10)" },
     { value = "raid20", name = "Raid (20)" },
     { value = "raid40", name = "Raid (40)" },